+++ sphinx/cmd/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+"""
+    sphinx.cmd
+    ~~~~~~~~~~
+
+    Modules for command line executables.
+
+    :copyright: Copyright 2007-2017 by the Sphinx team, see AUTHORS.
+    :license: BSD, see LICENSE for details.
+"""
